moved custom error lookup into createCommit
